+++ moban/core/context.py
@@ -1,7 +1,8 @@
 import os
 import copy
 
-from moban import utils, constants, exceptions
+from moban import constants, exceptions
+from moban.core import utils
 from moban.externals import reporter
 from moban.program_options import OPTIONS
 from moban.core.data_loader import merge, load_data
