+++ moban/deprecated/repo.py
@@ -2,8 +2,8 @@ import sys
 import subprocess
 
 from moban import constants, exceptions
-from moban.utils import mkdir_p
 from moban.externals import reporter, file_system
+from moban.core.utils import mkdir_p
 
 
 def git_clone(requires):
