+++ moban/mobanfile/__init__.py
@@ -6,8 +6,8 @@ from collections import OrderedDict
 
 from moban import core, constants
 from lml.utils import do_import
-from moban.utils import verify_the_existence_of_directories
 from moban.externals import reporter
+from moban.core.utils import verify_the_existence_of_directories
 from moban.deprecated import handle_copy, handle_requires
 from moban.core.data_loader import merge
 from moban.mobanfile.targets import (
